\name{gtkDfEditSetColumnName}
\alias{gtkDfEditSetColumnName}
\title{Set the columns of the RGtk2DfEdit object...}
\usage{gtkDfEditSetColumnName(object, idx, new.name)}
\description{Set the columns of the RGtk2DfEdit object}
\value{Returns the column names for the current object}
\arguments{\item{object}{The RGtk2DfEdit object}
\item{idx}{The col index}
\item{new.name}{The new column name}}
